+++ CHANGES.md
@@ -1,6 +1,7 @@
 ### Last Version
 
   - Change `assert\required` argument `$schema` as optional.
+  - Add `assert\optional`.
 
 ### 0.4.0 (2019-02-16)